MVP deployment setup
This is far from production ready and best practices, but is a step
towards an easy way to run a shared instance.

The app is served publicly, and the network connection to the database
is working.

The 3001 websocket traffic is not yet working, and I haven't attempted
to connect the actual application to the db.