Finished Banner, changed background and font size + weight, favicon and logo
### Home Banner
Finished and ready to be merged! Looks slick and modern

### Background
Removed the old background and made a new one that's a mix of Rhett's blue idea and AW1 (AWBW's) background.

### Font Weight and size
Tweaked font sizes and weights, these will work much better and include a wide array of weigths so we can make our fonts bolder or ligther depending on what we need.

### Favicon and logo
Deleted the advance wars logo and put our new logo, added a favicon ico that looks smoooove.