Add ThreeLinesText component, tweak primary color, do small tweaks to home component
### .button class
the .button class (not <button>) now properly styles the object to look like a button, even if its a link/<a>, it will still look like a button.

### ThreeLinesText.tsx
Small component that allows the developer to put a quick 3 lines of text with buttons!

### tailwind.config.js
Tweaked the orange to be brigther so it contrasts more with our dark background.

### layout.tsx
Now we use React.ReactElement because JSX.Element is outdated.