Add new color fixes to different components
Made the navbar use the new bg-primary theme instead of tailwind's default slate/blue.

Also added a new primary-dark, primary will refer mostly to the orange while secondary will be our green. This doesnt apply to bg-primary and co because those are bg colors, not the primary and secondary main colors if that makes sense.