Reorganized files and folders for better clarity (#147)
* Reorganized files and folders for better clarity

Reorganized the pages folder to be more streamlined, only pages that
have dynamic content (such as match/1 or news/patch-1.1) will need a
folder. Other pages can go by themelves without a folder and next.js
will display them correctly.

Home Page
the home page now has a BasicHome and a LoggedHome file, this is because
when the user is logged in, the home page will change to become a player
hub of sorts since logged in players wont visit the homepage since they
already know the site.

sass
Renamed some files and tidied up the naming conventions used.

Also deleted the layout component from pages and made it a default on
every page. This is because every single page will always have the
navbar and footer. Currently I cant think of a page that merits not
having either of these.

* Fixed formatting; Rerouted /howtoplay to /learn

* Git action fix

* Revert "Git action fix"

This reverts commit 710b200f4e9789aefe6b681e7b7ea18b4e2f7597.

* Fix incorrect usage of export default

---------

Co-authored-by: Ka Hung Lee <lee.kahung414@gmail.com>