Complete Markdown Article Render
Now we can render markdown into full blown articles with an index of
contents for each header. This also includes images as well.

From this markdown, we can also work to make the News and How to play
pages have examples that are more "real" since with the metadata in the
markdown can we work towards better examples.