Render Markdown for How to play guides and future news articles/balance patches (#162)
* Render markdown correctly

Right now the articles/flak route renders a simple placeholder markdown.
Now all that's left is to simply grab the markdown correctly and input
styles on top of it.

* Complete Markdown Article Render

Now we can render markdown into full blown articles with an index of
contents for each header. This also includes images as well.

From this markdown, we can also work to make the News and How to play
pages have examples that are more "real" since with the metadata in the
markdown can we work towards better examples.

* Fix ESLint problems

Basically added some ignore tags, tried fixing the type-safety issues
but it seems like a rabbithole im not willing to invest more time into.
Maybe Function or someone else with more experience can help.