Fixed minor issues with menu events and better removal
Now menu will only dissapear if we click on another tile, instead of
only doing so if we hover out. This makes it a much better experience
and it was achieved by simply creating a huge invisible rectangle behind
menu that, if its clicked, it removes the menu and the rectangle (rather
than needing to give an event listener to every other tile).

Also tweaked some minor style issues with the menu such as animations or
it not covering the factory anymore.