Plus minus buttons, center game-map (#181)
* Fixed minor issues with menu events and better removal

Now menu will only dissapear if we click on another tile, instead of
only doing so if we hover out. This makes it a much better experience
and it was achieved by simply creating a huge invisible rectangle behind
menu that, if its clicked, it removes the menu and the rectangle (rather
than needing to give an event listener to every other tile).

Also tweaked some minor style issues with the menu such as animations or
it not covering the factory anymore.

* Menus now appear correctly on the y-axis

Menus now dont ever go below the map, so even if your factory is at the
bottom of the map, the menu doesnt display below it but above it in a
very clean fashion.

* Add plus and minus buttons that work

Added a green (plus) and red (minus) squares that either increase or
decrease the scale of the map. This perfectly (or almost perfectly)
increases size without compromising our sprites looking worse/having
abberrations.

Also made sure to add these in a separate file, so our main file isn't
filled with too much stuff.

* Map displays correctly in center, buttons styled

Now the map correctly displays itself in the center correctly and scales
correctly from the center.

Plus and minus buttons also work correctly mow

* Fix menu issues with air/sea units, add x-axis check

Now the map can correctly display all the menus and never have any of
them leaving the game-map. Ensuring the map always stays map size.

* Deleted unused component, fix other minor issues

Deleted a component that would create a plus and minus button on pixi,
this is now handled outside of pixi.

Als

* hotfix code breaking issue

Not sure why but I deleted an if check and it breaks our map rendering
quite well. So Im going to let it be.