Ability to switch CO
This commit creates the initial component for our matches (I know we had
another one but I felt it would be faster just to re-do it on css
grid). Now when a match is created, we will see our player id, the map
name and the fact that its in setup.

This commit also has eliminated some replicated co images and replaced
them accordingly.

I also created my first tRPC procedure, "switchCO" which allows us to
switch COs for a match. However, all of this still needs polish and
cleanup!