Able to join other games effectively
Now we can join other games we are not part of and games are displayed
effectively. Showing the difference between games you are part of VS
games you are not part of and therefore can join. Also, now if you
belong to a match, you are always displayed as the first player no
matter what.