Polish match setup, show COs on match
With this update, the whole match card component is functional and
working well. When opening a game, now we also see both COs, and when we
ready up, the match goes to live and our CO pops up a little bit (but
does become greyed out when not ready).

This sets the floor for alot of the things needed to start working on
game development!