Match creation, setup, joining and spectate (#183)
* Show other user matches, make maps tighter

Im just getting my feet wet with tRPC and our backend, but I've managed
to make the your-matches page not just show all matches but divide
between matches a user created and those that the user did not create.

This is so we can start the whole process of turns which is necessary to
get us to be able to create units lol.

* Ability to switch CO

This commit creates the initial component for our matches (I know we had
another one but I felt it would be faster just to re-do it on css
grid). Now when a match is created, we will see our player id, the map
name and the fact that its in setup.

This commit also has eliminated some replicated co images and replaced
them accordingly.

I also created my first tRPC procedure, "switchCO" which allows us to
switch COs for a match. However, all of this still needs polish and
cleanup!

* Able to join other games effectively

Now we can join other games we are not part of and games are displayed
effectively. Showing the difference between games you are part of VS
games you are not part of and therefore can join. Also, now if you
belong to a match, you are always displayed as the first player no
matter what.

* Polish match setup, show COs on match

With this update, the whole match card component is functional and
working well. When opening a game, now we also see both COs, and when we
ready up, the match goes to live and our CO pops up a little bit (but
does become greyed out when not ready).

This sets the floor for alot of the things needed to start working on
game development!

* Leaderboard Fix

Fixed the issue where the nations images will not show up properly, also changed the enums to use zod schemas and types instead.

* Undefined Handlers V2

Make it so you don't need to manually add a playerId to localstorage to be able to enter the page.

* Lint Fix

* Current Player gets to be default on the select

* Loading State for Players Dropbox

When there are no players loaded the page will show a text of loading players, so for the future we can expand that.

* Nation Icons fixes

Leaderboards page will render the nation icons correctly.

* No PlayerId in LS fix

Fix the error where the players dropdown will never show up despite the fact the owned players are already loaded, now it should work correctly.

---------

Co-authored-by: Enrique <enriqueachavez@hotmail.com>
Co-authored-by: Enrique Chávez <56766746+CliveGlitch@users.noreply.github.com>