NEW: Next Auth custom adapter
The adapter makes Next Auth able to save users into the database when the user logins for the first time with any provider that is not credentials (like discord), it has more applications but for now that is what it can do properly.

I've made a custom adapter so we can add more logic when adding users or any other action like sessions, in this example I made a default player every time it creates an user.