Update Seed Match
The seed match we had was missing some properties (teamMapping,
labUnitTypes), causing errors. The
new seed has the correct properties now.

Also, apparently match.units is undefined, which results in an error. I
made a small if check that if its undefined, we return an empty array (I
assume no units). a TODO was placed bringing this up, since this doesn't
seem to be intended behavior.

Also placed a small comment on match-rules teamMapping (yes this comment
is necessary, I had to ping you guys to understand what it was, so its
not clear enough).