Match Creation Possible
Now Matches can be created and displayed correctly.

Most changes are in the frontend, updating to the new COID schema which
has a name and version. Also updating how we send data to create
matches and the capability to switch CO, army and status (ready or not
ready)

However, some changes were done in the backend to create.ts because it
was unable to find the mapId since it was not included in the input
(mapMiddleware would throw the error, and logging the input would show
mapId missing, which in the frontend would also show an error that mapId
is not a necessary property or something along those lines).

A big issue right now is that while we can switch COs and armies, these
changes do not persist into the database. We need to do stuff like
INSERT and COMMIT with prisma.