Match Update Persist, join, leave WIP
In this update, now changes to the match persist to the database (can
see through command "prisma studio").

Currently, creating or joining  a match, changing CO, army and ReadyState all work correctly. However, leaving and starting the match have their own issues.

Leaving a match
Leaving a match effectively deletes it from the database but does not
dissapear in the FrontEnd, the match appears as a ghost until some
action makes the frontend fully re-fetch everything.

Starting a match
Right now the app does not handle well the match-start event and it ends
up crashing the database hard. Only way to fix is by resetting the
database afaik.

Also for the prisma changes, this is all my first time working with
prisma directly so if you see some atrocious error or issue. I'll be
happy to take the feedback! Just bear in mind I'm learning.