Revamped articles
All of the changes here have been done to make the creation of articles as easy as possible through markdown files.
All the articles listed in the /news page are now all pulled from src/frontend/utils/articles.

- The file name of the markdown files are now used as the title for the articles.
The only issue I see with this is that articles with titles such as "this is omg!!" and "this is omg!" are seen the same. In other words, when we click on them, they both redirect to news/this-is-omg.

- Replaced subtitle property with description.
To make the articles SEO friendly, I added the description to <meta name="description" content={HERE}/> in src/pages/news/[id].tsx. The property name 'subtitle' didn't make sense.

Potential issue:
- getArticleData:
When accessing an article page, the only way the website can tell which markdown file to render is by comparing the slug equivalent of the file names to the slug in the url.
For example, when loading up the article for news/good-girl-lash, it loops through all the markdown files, converting each file name to slug form (ex. "Welcome back!" -> "welcome-back").
Each slug is compared with "good-girl-lash".
If the same, then that is the article it is going to render.
I am not sure if this is particularly efficient.