Fixed player slot issue
There was an issue when the scenario below is done:
1. Player 1 creates match
2. player 2 joins match
3. player 1 leaves match
4. a player joins the same match

This was due to the fact that players that join matches join specifically on playerSlot 1.
In the scenario above in step 4, that player tries to occupy playerSlot 1 when player 2 is already there.

This fix solves this scenario by seeing which slot is available to join, albiet superficially.

I will be working on this.