Enhance map rendering (#216)
* Bless CliveG

* Refactor code and enhance map rendering

This commit refactors the code in `map-render.ts` and `[matchId].tsx` for improved clarity and efficiency. The key changes include enabling zIndex on mapContainer children for better control and switching from local constants to using useRef for managing the PIXI application and mapContainer. Several changes directly enhance the map rendering process by providing a cleaner setup and management of the PIXI application and mapContainer within a useEffect.