Fixed MatchCard issue when switching between two player of the same match
Given that two players only have one match each, in which their match is only a 1v1 between them, the CO and country banner color on the left part of the MatchCard would display incorrectly.

For example,
Player 1 has one match
Player 2 has one match
There is a 1v1 match between player 1 and player 2.
When current player is player 1, the left CO and country is Andy and Orange Star, respectively and the right CO and country is Flak and Black Hole, respectively.
When setting the current player to 2, the left CO and country would still be Andy and Orange Star.

This is an issue because the left player is the current player (player 2) and is supposed to be Flak and Black Hole.

This scenario described might be a bit specific where the issue might have been more general.

By using useEffect, this issue is fixed. Reach out to me for more details.