Fixed issue where matches are not displaying correct information when switching players in your-matches page.
When updating the CO or country of one player of a match and then switching to another player, the changes would not be reflected on the page.
A refresh would be needed to see the update because the CO/country updates in the database, but not from the client's perspective.

By refetching the matches after changing players, this issue is fixed.