Switching to the same army/slot as the opponent is now not possible
To do:
- Make sure that when a player joins a match, the nation they're assigned to is not the same as the player already in the match.
Currently, when a player joins a match, the nation they're assigned to is hard-coded to blue-moon.
- Make the backend enforce against players being in the same nation/slot
From the UI, it is impossible for a player to select a nation that is the same as another player due to the UI not allowing it.
However, even if a call to the backend is done with the wrong nation, there is nothing stopping the change from happening in the match router.

I will be addressing these issues in my next commit. Please feel free to contact me for details.