No more collisions between players having the same slots/armies
When a player joins a game, it no longer defaults to blue moon. Instead, it picks a random available country for the player.

The backend now checks if a country a player attempts to switch to is occupied by another player.
If so, it throws an error.
Same deal for player slots.

Changes to existing code:
- switchCo, switchArmy, and switchSlots have been consolidated into switchOptions due to identical code.