BuildMenu uses funds
This update is mostly about having the build menu correctly use funds to
display units we can and can't afford. It also makes it so only the
player with the turn can prop the build menu since any other user/player
should not be able to build if its not their turn.