Build Event and Code Improvements
If you try building an unit, an event is triggered! This shows on the db
but for now, the frontend is struggling to put the events properly into
memory.

There we also some minor improvements to errors when requesting a
playerId and reduction of code needed to find which player has the
current turn.