Refactor SpriteSheet and Use other country colors
The way we use spritesheets has been refactored. Now instead of being an
array, its an object such as "black-hole" and that has the bh sprites.
So we can just call spritesheet["black-hole"] rather than having to
remember which index in the array that is.

I've also polished our types when it comes to loading and getting
spritesheets so its cleaner.

Also now we correctly render countries based on color and player slot.
So if playerSlot 0 is black hole, all their properties will correctly
render as black-hole.