Fixed type errors associated with buying units and tests
I basically just wrapped actionMutation.mutateAsync with a function so that passing it down props won't make typescript scream at us.

Also, edited the test so that it is up-to-date with all the changes. (I am not sure if it works as intended)

Extra:
user is now in the dependency array of the useEffect in players.tsx