Finished implementation of in game chat boxes (#223)
Within each ongoing match, a fixed chat box now appears on the right side.
The messages are sent and received in real time using the same subscription hook for match events.

Notably, the chat messages are not stored in local memory in the server.
Every time a player sends a message, it gets created in the database and then an event is broadcasted.
I wanted to keep match-wrapper as simple as it can be.