Clearly separated emittables and events
In game chat messages calls for the separation of the match actions broadcasted and other actions broadcasted.
Match actions need to be stored as events for match replays. In contrast, other actions, such as sending messages in a live match chat box,
are not stored as events.

To address this, emittable and events are clearly defined in src/shared/types/events.ts.
I will be renaming this file in the next commit.