Units can now Move + submenu
Now units can move! Plus the submenu works as intended.

Moving units still needs to be polished (can't wait on same spot atm...)
but they move pretty well!

Albeit the work is not done yet, im pushing this code to avoid creating
a humongous commit later on (which this one is already pretty big)

Nonetheless, this is functional and works good enough to be commited!