Interactable tiles fixes
Minor fixes of some interactable tiles such as making sure menus would
be in the corrext x-y axis. Also making attackable units produce a
correct render of the battle forecast/probabilities.

Also minor bugfix of units not being able to do an action in place
(having an error on units not moving and doing an action).